Re organize and modularize praser
