+++ stu.cabal
@@ -26,7 +26,8 @@ library
         OverloadedStrings, LambdaCase
 
     exposed-modules:
-        Parser, AST, CodeGen, Analysis, Types 
+        Parser, AST, CodeGen, Analysis, Types,
+        Parser.Types, Parser.Expr, Parser.Util 
 
     -- other-extensions:
     build-depends:
@@ -62,7 +63,7 @@ executable stu
 
     -- Modules included in this executable, other than Main.
     other-modules:
-        Parser, AST, CodeGen, Analysis, Types 
+        Parser, AST, CodeGen, Analysis, Types       
 
     -- LANGUAGE extensions used by modules in this package.
     -- other-extensions:
